+++ app/app/globals.h
@@ -2,11 +2,7 @@
 #ifndef BASALT_APP_GLOBALS_H
 #define BASALT_APP_GLOBALS_H
 
-#include "runtime/platform/Platform.h"
-
 #include "runtime/shared/Config.h"
-
-#include <vector>
 
 namespace basalt {
 
@@ -19,8 +15,6 @@ struct View;
 namespace win32 {
 
 extern WindowMode sWindowMode;
-
-extern std::vector<platform::PlatformEventCallback> sEventListener;
 
 extern gfx::View sCurrentView;